extension: Add a start-if-stopped (server) command
Handy for, e.g., another extension such as [1], if willing to ask this
one to start the trace server (if known as stopped). Say, right before
opening a trace using [1], or the like. This command is also available
as usual through Command Palette.

[1] https://github.com/eclipse-cdt-cloud/vscode-trace-extension

Make sure to un-define the previously started server right after killed,
if/while having been stopped. That mostly enables this new command.

Signed-off-by: Marco Miller <marco.miller@ericsson.com>