extension: Store updated server pid in workspace
Rely on this pid state now to be able to always stop the previously
started trace server. Meaning even if recovering upon some reload or
restart of the extension.

Before this change, this.server{.pid} was lost after such a recovery
event. This prevented the extension from being able to honor a stop
(trace server) command, for an instance started before that recovery.

Add the hereby necessary "or reset" aspect to the stop command. Make the
act of internally resetting the known pid explicit, this way. Using the
stop command upon a trace server that was already stopped outside of the
extension is resetting that pid state now.

Fixes #11. Upon merging this very PR, the related #1's README warning
[1] shall be removed, as it gets addressed herein.

[1]https://github.com/marco-miller/vscode-trace-server/pull/1/files#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R90

Signed-off-by: Marco Miller <marco.miller@ericsson.com>