TraceServer: Add an explicit foreign server check
If a foreign server is already running, the start command used to miss
that and keep going, upon a yet defined server.pid. Add an isUp check
now to prevent startup from proceeding, should a known foreign server be
already running.

Checking for server.exitCode or server.signalCode at this early startup
stage couldn't fix this, because both are prematurely set as 'none',
which means a running spawned server.

Contributes to fixing #8.

Signed-off-by: Marco Miller <marco.miller@ericsson.com>