TraceServer: Remove extra server-exit stopOrReset
Before this fix, an extra stopOrReset was performed for nothing, upon
stopping (thus exiting) the previously started server. This caused an
extra (confusing) warning message being showed which was unrelated
anyway. The latter was about not stopping, but this use case is truly
about restarting the server.

Contributes to fixing #8.

Signed-off-by: Marco Miller <marco.miller@ericsson.com>