TraceServer: Fix case of starting upon stored pid
Before this fix, a warning was always showed to the user upon an already
"running" server, even in the case of a no longer running server.
Meaning, in the case where a pid is still stored despite the
corresponding process(es) not running anymore.

Do an internal reset followed by a start attempt in such a case, to
honor the genuine start command while duly resetting internally.

Contributes to fixing #8.

Signed-off-by: Marco Miller <marco.miller@ericsson.com>